Fix User model to_json url value
The User model’s to_json() returned an object whose url value was for a
post whose id matched the user’s; this wasn’t the intended behavior, so
this commit changes the url value to be the intended resource for the
user.  Included is a test to check both the keys that should be present
in a user’s json representation as well as a check that the url
returned is for a user.